Minor bugs fixed and warnings removed

+++ Utils/stringUtils.c
@@ -1,6 +1,6 @@
 #include "stringUtils.h"
 
-DWORD getNumOfSubstrings(const TCHAR* str, const TCHAR* substr, DWORD substrLength) {
+DWORD getNumOfSubstrings(const TCHAR* str, const TCHAR* substr, size_t substrLength) {
 	TCHAR* temp = (TCHAR*)str;
 	TCHAR* curSubstrPos = NULL;
 	DWORD counter;
@@ -15,19 +15,19 @@ const TCHAR* strReplace(const TCHAR* srcStr, const TCHAR* strToReplace, const TC
 		// TODO: error message
 		return FALSE;
 	}
-	DWORD offsetInSrcStr = 0;
+	size_t offsetInSrcStr = 0;
 
-	DWORD strToReplaceLength = wcslen(strToReplace);
-	DWORD replaceWithLength = wcslen(replaceWith);
-	DWORD srcStrLength = wcslen(srcStr);
+	size_t strToReplaceLength = wcslen(strToReplace);
+	size_t replaceWithLength = wcslen(replaceWith);
+	size_t srcStrLength = wcslen(srcStr);
 	
 	DWORD numOfSubstrings = getNumOfSubstrings(srcStr, strToReplace, strToReplaceLength);
 	if (0 == numOfSubstrings) {
 		return srcStr;
 	}
 
-	DWORD outStrElems = srcStrLength + numOfSubstrings * (replaceWithLength - strToReplaceLength) + 1;
-	DWORD outStrLength = sizeof(TCHAR) * outStrElems;
+	size_t outStrElems = srcStrLength + numOfSubstrings * (replaceWithLength - strToReplaceLength) + 1;
+	size_t outStrLength = sizeof(TCHAR) * outStrElems;
 	TCHAR* outStr = (TCHAR*)malloc(outStrLength);
 
 	TCHAR* strToReplacePtr = NULL;
@@ -35,7 +35,7 @@ const TCHAR* strReplace(const TCHAR* srcStr, const TCHAR* strToReplace, const TC
 	
 	while (numOfSubstrings--) {
 		strToReplacePtr = wcsstr(srcStr + offsetInSrcStr, strToReplace);
-		DWORD indexInStr = strToReplacePtr - srcStr;
+		DWORD indexInStr = (DWORD)(strToReplacePtr - srcStr);
 		if (0 != indexInStr) {
 			wcsncpy_s(temp, outStrElems, srcStr, indexInStr);
 			temp += indexInStr;
@@ -47,7 +47,7 @@ const TCHAR* strReplace(const TCHAR* srcStr, const TCHAR* strToReplace, const TC
 		offsetInSrcStr += indexInStr + strToReplaceLength;
 	}
 	if (offsetInSrcStr != srcStrLength) {
-		DWORD restOfStrLen = wcslen(srcStr + offsetInSrcStr) + 1;
+		size_t restOfStrLen = wcslen(srcStr + offsetInSrcStr) + 1;
 		wcscpy_s(temp, restOfStrLen, srcStr + offsetInSrcStr);
 	}
